More singly Linked Lists

+++ 0x13-more_singly_linked_lists/0-print_listint.c
@@ -1,18 +1,16 @@
 #include <stdio.h>
 #include "lists.h"
-#include <stdlib.h>
 /**
  * print_listint - prints all the element of a node list
  *
- * @h - new listint_t
+ * @h - pointer to a list
  *
- * Return: size_t
+ * Return: int
  */
-
 size_t print_listint(const listint_t *h)
 {
 	const listint_t *new = h;
-	size_t count = 0;
+	int count = 0;
 
 	while (new != NULL)
 	{
